- moved dialog components to dialogs folder
