+++ coops-ui/src/components/dialogs/AccountAuthorizationDialog.vue
