+++ coops-ui/src/components/dialogs/CoaWarningDialog.vue
