+++ coops-ui/src/components/dialogs/ConfirmDialog.vue
