+++ coops-ui/src/components/dialogs/DashboardUnavailableDialog.vue
