+++ coops-ui/src/components/dialogs/DeleteErrorDialog.vue
