+++ coops-ui/src/components/dialogs/DownloadErrorDialog.vue
