+++ coops-ui/src/components/dialogs/ResumeErrorDialog.vue
