+++ coops-ui/src/components/dialogs/index.ts
@@ -1,0 +1,21 @@
+import AccountAuthorizationDialog from './AccountAuthorizationDialog.vue'
+import CoaWarningDialog from './CoaWarningDialog.vue'
+import ConfirmDialog from './ConfirmDialog.vue'
+import DashboardUnavailableDialog from './DashboardUnavailableDialog.vue'
+import DeleteErrorDialog from './DeleteErrorDialog.vue'
+import DownloadErrorDialog from './DownloadErrorDialog.vue'
+import PaymentErrorDialog from './PaymentErrorDialog.vue'
+import ResumeErrorDialog from './ResumeErrorDialog.vue'
+import SaveErrorDialog from './SaveErrorDialog.vue'
+
+export {
+  AccountAuthorizationDialog,
+  CoaWarningDialog,
+  ConfirmDialog,
+  DashboardUnavailableDialog,
+  DeleteErrorDialog,
+  DownloadErrorDialog,
+  PaymentErrorDialog,
+  ResumeErrorDialog,
+  SaveErrorDialog
+}
