+++ coops-ui/src/views/StandaloneDirectorsFiling.vue
@@ -289,10 +289,7 @@ import SbcFeeSummary from 'sbc-common-components/src/components/SbcFeeSummary.vu
 import { SummaryDirectors, SummaryCertify, SummaryStaffPayment } from '@/components/common'
 
 // Dialog Components
-import ConfirmDialog from '@/components/ConfirmDialog.vue'
-import PaymentErrorDialog from '@/components/AnnualReport/PaymentErrorDialog.vue'
-import ResumeErrorDialog from '@/components/AnnualReport/ResumeErrorDialog.vue'
-import SaveErrorDialog from '@/components/AnnualReport/SaveErrorDialog.vue'
+import { ConfirmDialog, PaymentErrorDialog, ResumeErrorDialog, SaveErrorDialog } from '@/components/dialogs'
 
 // Mixins
 import { EntityFilterMixin } from '@/mixins'
