+++ coops-ui/src/views/StandaloneOfficeAddressFiling.vue
@@ -168,10 +168,7 @@ import axios from '@/axios-auth'
 import { mapState, mapGetters } from 'vuex'
 
 // Dialogs
-import ConfirmDialog from '@/components/ConfirmDialog.vue'
-import PaymentErrorDialog from '@/components/AnnualReport/PaymentErrorDialog.vue'
-import ResumeErrorDialog from '@/components/AnnualReport/ResumeErrorDialog.vue'
-import SaveErrorDialog from '@/components/AnnualReport/SaveErrorDialog.vue'
+import { ConfirmDialog, PaymentErrorDialog, ResumeErrorDialog, SaveErrorDialog } from '@/components/dialogs'
 
 // Components
 import { OfficeAddresses } from '@/components/common'
